LayoutTests/media/modern-media-controls/media-controls/media-controls-constructor.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/scheduler.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/layout-node.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/layout-item.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/button.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/icon-service.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/start-button.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/icon-button.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/play-pause-button.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/skip-back-button.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/airplay-button.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/pip-button.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/fullscreen-button.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/slider.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/scrubber.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/time-label.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/time-control.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/placard.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/airplay-placard.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/pip-placard.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/media-controls.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

description("Testing the <code>MediaControls</code> constructor.");

iconService.directoryPath = "../../../../Source/WebCore/Modules/modern-media-controls/images";

const mediaControls = new MediaControls({ layoutTraits: LayoutTraits.macOS });

shouldBeEqualToString("mediaControls.element.localName", "div");
shouldBeEqualToString("mediaControls.element.className", "media-controls");

debug("");
shouldBe("mediaControls.width", "300");
shouldBe("mediaControls.height", "150");
shouldBe("mediaControls.layoutTraits", "LayoutTraits.macOS");
shouldBeFalse("mediaControls.showsStartButton");

debug("");
shouldBeTrue("mediaControls.startButton instanceof StartButton");
shouldBeTrue("mediaControls.playPauseButton instanceof PlayPauseButton");
shouldBeTrue("mediaControls.skipBackButton instanceof SkipBackButton");
shouldBeTrue("mediaControls.airplayButton instanceof AirplayButton");
shouldBeTrue("mediaControls.pipButton instanceof PiPButton");
shouldBeTrue("mediaControls.fullscreenButton instanceof FullscreenButton");
shouldBeTrue("mediaControls.timeControl instanceof TimeControl");
shouldBeTrue("mediaControls.airplayPlacard instanceof AirplayPlacard");
shouldBeTrue("mediaControls.pipPlacard instanceof PiPPlacard");

debug("");
shouldBeTrue("mediaControls.controlsBar instanceof LayoutItem");
shouldBeEqualToString("mediaControls.controlsBar.element.localName", "div");
shouldBeEqualToString("mediaControls.controlsBar.element.className", "controls-bar");

debug("");
const mediaControlsWithArguments = new MediaControls({ width: 600, height: 300, layoutTraits: LayoutTraits.macOS });
shouldBe("mediaControlsWithArguments.width", "600");
shouldBe("mediaControlsWithArguments.height", "300");

debug("");

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
